Network.PFQ

Description

The PFQ module is a low level binding to the functions in libpfq. See https://github.com/pfq/PFQ/wiki for more information.

Types

data PFqTag

Packet capture handle.

data Statistics

PFQ statistics.

Constructors

Statistics

Fields

sReceived :: Integer

packets received

sLost :: Integer

packets lost

sDropped :: Integer

packets dropped

sSent :: Integer

packets sent

sDiscard :: Integer

packets discarded

sFailure :: Integer

packets Tx failure

sForward :: Integer

packets forwarded to devices

sKernel :: Integer

packets forwarded to kernel

data NetQueue

Capture Queue handle.

Constructors

NetQueue

Fields

qPtr :: Ptr PktHdr

pointer to the memory mapped queue

qLen :: !Word64

queue length

qSlotSize :: !Word64

size of a slot = pfq header + packet

qIndex :: !Word32

index of the queue

data Packet

Descriptor of the packet.

Constructors

Packet

Fields

pHdr :: Ptr PktHdr

pointer to pfq packet header

pData :: Ptr Word8

pointer to the packet data

pIndex :: !Word32

index of the queue

data PktHdr

PFQ packet header.

Constructors

PktHdr

Fields

hMark :: !Word32

skb 32-bits mark

hState :: !Word32

opaque 32-bits state

hSec :: !Word32

timestamp (seconds)

hNsec :: !Word32

timestamp (nanoseconds)

hIfIndex :: !Word32

interface index

hGid :: !Word32

group id

hLen :: !Word16

packet length (wire size)

hCapLen :: !Word16

capture length

hTci :: !Word16

vlan tci

hHwQueue :: !Word8

hardware queue index

hCommit :: !Word8

commit bit

type Callback = PktHdr -> Ptr Word8 -> IO ()

The type of the callback function passed to dispatch.

newtype ClassMask

ClassMask type.

Constructors

ClassMask

Fields

getClassMask :: CULong

class_default :: ClassMask

class_user_plane :: ClassMask

class_control_plane :: ClassMask

class_control :: ClassMask

class_any :: ClassMask

newtype GroupPolicy

Group policy type.

Constructors

GroupPolicy

Fields

getGroupPolicy :: CInt

policy_undefined :: GroupPolicy

policy_priv :: GroupPolicy

policy_restricted :: GroupPolicy

policy_shared :: GroupPolicy

newtype PFqConstant

Generic pfq constant.

Constructors

PFqConstant

Fields

getConstant :: Int

any_device :: PFqConstant

any_queue :: PFqConstant

any_group :: PFqConstant

version_code :: PFqConstant

major_version :: PFqConstant

minor_version :: PFqConstant

patchlevel_version :: PFqConstant

version :: String

data SocketParams

SocketParams data type.

To be passed as argument to openParam function.

Constructors

SocketParams

Fields

parCaplen :: Int

capture len

parRxSlots :: Int

socket Rx queue length

parTxSlots :: Int

socket Tx queue length

parPolicy :: GroupPolicy

default group policy: policy_undefined means no group

parClass :: ClassMask

socket class

defaultSocketParams :: SocketParams

Default values of Socket parameters.

Socket and Groups

open

Arguments

:: Int
caplen

-> Int
number of Rx slots

-> Int
number of Tx slots

-> IO (ForeignPtr PFqTag)

Open a socket and create a new private group.

The default values for class mask and group policy are class_default and policy_priv, respectively.

openNoGroup

Arguments

:: Int
caplen

-> Int
number of Rx slots

-> Int
number of Tx slots

-> IO (ForeignPtr PFqTag)

Open a socket. No group is joined or created.

Groups can later be joined by means of joinGroup function.

openGroup

Arguments

:: ClassMask
ClassMask (e.g., class_default mappend class_control_plane)

-> GroupPolicy
policy for the group

-> Int
caplen

-> Int
number of Rx slots

-> Int
number of Tx slots

-> IO (ForeignPtr PFqTag)

Open the socket and create a new group with the specified parameters.

If group_policy is policy_undefined no gorup is joined or created.

openParam

Arguments

:: SocketParams
parameters

-> IO (ForeignPtr PFqTag)

Open the socket with SocketParams.

Default values are defined as defaultSocketParams.

close :: Ptr PFqTag -> IO ()

Close the socket.

Release the shared memory, stop kernel threads.

enable :: Ptr PFqTag -> IO ()

Enable the socket for packets capture and transmission.

Allocate the shared memory for socket queues, possibly using the Linux HugePages support. If the enviroment variable PFQ_HUGEPAGES is set to 0 (or PFQ_NO_HUGEPAGES is defined) standard 4K pages are used.

disable :: Ptr PFqTag -> IO ()

Disable the socket.

Release the shared memory, stop kernel threads.

getId :: Ptr PFqTag -> IO Int

Return the id of the socket.

getGroupId :: Ptr PFqTag -> IO Int

Return the group-id of the socket.

isEnabled :: Ptr PFqTag -> IO Bool

Check whether the socket is enabled.

Socket control

joinGroup

Arguments

:: Ptr PFqTag

-> Int
group id

-> ClassMask
class mask

-> GroupPolicy
group policy

-> IO ()

Join the group with the given class mask and group policy.

leaveGroup

Arguments

:: Ptr PFqTag

-> Int
group id

-> IO ()

Leave the group specified by the group id.

bind

Arguments

:: Ptr PFqTag

-> String
device name

-> Int
queue index (or any_queue constant)

-> IO ()

Bind the main group of the socket to the given device/queue.

bindGroup

Arguments

:: Ptr PFqTag

-> Int
group id

-> String
device name

-> Int
queue index (or any_queue constant)

-> IO ()

Bind the group to the given device/queue.

unbind

Arguments

:: Ptr PFqTag

-> String
device name

-> Int
queue index (or any_queue constant)

-> IO ()

Unbind the main group of the socket from the given device/queue.

unbindGroup

Arguments

:: Ptr PFqTag

-> Int
group id

-> String
device name

-> Int
queue index

-> IO ()

Unbind the group from the given device/queue.

egressBind

Arguments

:: Ptr PFqTag

-> String
device name

-> Int
queue index

-> IO ()

Set the socket as egress and bind it to the given device/queue.

The egress socket is be used by groups as network forwarder.

egressUnbind :: Ptr PFqTag -> IO ()

Unset the socket as egress.

Socket parameters

timestampingEnable

Arguments

:: Ptr PFqTag

-> Bool
toggle: True is on, False off.

-> IO ()

Enable/disable timestamping for packets.

isTimestampingEnabled :: Ptr PFqTag -> IO Bool

Check whether timestamping for packets is enabled.

setWeight

Arguments

:: Ptr PFqTag

-> Int
weight of socket (valid range is [1,16))

-> IO ()

Set the weight of the socket for the steering phase.

getWeight :: Ptr PFqTag -> IO Int

Return the weight of the socket.

setPromisc

Arguments

:: Ptr PFqTag

-> String
device name

-> Bool
toggle: True is on, False off.

-> IO ()

Set the promiscuous mode for the given interface.

getCaplen :: Ptr PFqTag -> IO Int

Return the capture length of packets, in bytes.

setCaplen

Arguments

:: Ptr PFqTag

-> Int
caplen (bytes)

-> IO ()

Specify the capture length of packets, in bytes.

Capture length must be set before the socket is enabled.

getRxSlots :: Ptr PFqTag -> IO Int

Return the length of the Rx queue, in number of packets.

setRxSlots

Arguments

:: Ptr PFqTag

-> Int
number of Rx slots

-> IO ()

Specify the length of the Rx queue, in number of packets.

getRxSlotSize :: Ptr PFqTag -> IO Int

Return the length of a Rx slot, in bytes.

getTxSlots :: Ptr PFqTag -> IO Int

Return the length of the Tx queue, in number of packets.

setTxSlots

Arguments

:: Ptr PFqTag

-> Int
number of Tx slots

-> IO ()

Specify the length of the Tx queue, in number of packets.

getMaxlen :: Ptr PFqTag -> IO Int

Return the max transmission length of packets, in bytes.

Packet capture

read

Arguments

:: Ptr PFqTag

-> Int
timeout (msec)

-> IO NetQueue

Read packets in place.

Wait for packets and return a NetQueue descriptor.

The memory of the socket queue is reset at the next read. A timeout is specified in microseconds.

dispatch

Arguments

:: Ptr PFqTag

-> Callback
packet processing function

-> Int
timeout (msec)

-> IO ()

Collect and process packets.

This function is passed a function Callback which is called on each packet.

getPackets :: NetQueue -> IO [Packet]

Return the list of Packet stored in the NetQueue.

getPacketHeader :: Packet -> IO PktHdr

Return the PktHdr of the given Packet.

isPacketReady :: Packet -> IO Bool

Check whether the Packet is ready or not.

waitForPacket :: Packet -> IO ()

Wait until the Packet is ready.

newtype VlanTag

Vlan tag.

Constructors

VlanTag

Fields

getVid :: CInt

vlan_untag :: VlanTag

vlan_anytag :: VlanTag

vlanFiltersEnable

Arguments

:: Ptr PFqTag

-> Int
group id

-> Bool
toggle: True is on, False off.

-> IO ()

Enable/disable vlan filtering for the given group.

vlanSetFilter

Arguments

:: Ptr PFqTag

-> Int
group id

-> VlanTag
vlan id

-> IO ()

Specify a capture vlan filter for the given group.

In addition to standard vlan ids, valid ids are also vlan_untag and vlan_anytag.

vlanResetFilter

Arguments

:: Ptr PFqTag

-> Int
group id

-> VlanTag
vlan id

-> IO ()

Reset the vlan filter for the given group.

Packet transmission

send

Arguments

:: Ptr PFqTag

-> ByteString
bytes of packet

-> Int
copies

-> Int
fhint

-> IO Bool

Store the packet and transmit the packets in the queue.

The queue is flushed every fhint packets. Requires the socket is bound for transmission to a net device and queue. See bindTx.

sendTo

Arguments

:: Ptr PFqTag

-> ByteString
bytes of packet

-> Int
ifindex

-> Int
qindex

-> Int
copies

-> Int
fhint

-> IO Bool

Store the packet and transmit the packets in the queue.

The queue is flushed every fhint packets.

sendAsync

Arguments

:: Ptr PFqTag

-> ByteString
bytes of packet

-> Int
copies

-> IO Bool

Transmit the packet asynchronously.

The transmission is handled by PFQ kernel threads. Requires the socket is bound for transmission to one (or multiple) PFQ kernel threads. See bindTx.

sendAt

Arguments

:: Ptr PFqTag

-> ByteString
bytes of packet

-> TimeSpec
active timestamp

-> Int
copies

-> IO Bool

Transmit the packet asynchronously.

The transmission takes place asynchronously at the given Timespec time. Requires the socket is bound for transmission to one (or multiple) PFQ kernel threads. See bindTx.

transmitQueue

Arguments

:: Ptr PFqTag

-> Int
queue index (0 is sync tx queue)

-> IO ()

Flush the Tx queue(s)

Transmit the packets in the Tx queues of the socket.

bindTx

Arguments

:: Ptr PFqTag

-> String
device name

-> Int
hw queue index

-> Int
PFQ thread id (number)

-> IO ()

Bind the socket for transmission to the given device name and queue.

unbindTx :: Ptr PFqTag -> IO ()

Unbind the socket for transmission.

Unbind the socket for transmission from any device/queue.

pfq-lang

setGroupComputation

Arguments

:: Ptr PFqTag

-> Int
group id

-> Function (Qbuff -> Action Qbuff)
expression (pfq-Lang)

-> IO ()

Specify a functional computation for the given group.

The functional computation is specified as a pfq-lang expression.

setGroupComputationFromString

Arguments

:: Ptr PFqTag

-> Int
group id

-> String
pfq-lang expression

-> IO ()

Specify a functional computation for the given group, as pfq-lang program from string.

setGroupComputationFromFile

Arguments

:: Ptr PFqTag

-> Int
group id

-> FilePath
pfq-lang file

-> IO ()

Specify a functional computation for the given group, as pfq-lang program from file.

setGroupComputationFromDescr

Arguments

:: Ptr PFqTag

-> Int
group id

-> [FunctionDescr]
expression (pfq-lang)

-> IO ()

Specify a functional computation for the given group.

The functional computation is specified as a list of FuncitonDescr.

setGroupComputationFromJSON

Arguments

:: Ptr PFqTag

-> Int
group id

-> String
decode (json) :: [FunctionDesc]

-> IO ()

Specify a functional computation for the given group, from JSON description.

Statistics and counters

getStats :: Ptr PFqTag -> IO Statistics

Return the socket statistics.

getGroupStats

Arguments

:: Ptr PFqTag

-> Int
group id

-> IO Statistics

Return the statistics of the given group.

getGroupCounters

Arguments

:: Ptr PFqTag

-> Int
group id

-> IO Counters

Return the set of counters of the given group.