Network.PFQ.Lang.Experimental

Experimental Functions

This set of experimental functions may be subject to changes in future releases

dummy :: Int -> NetFunction

dummy_ip :: IPv4 -> NetFunction

dummy_vector :: [Int] -> NetFunction

dummy_string :: String -> NetFunction

dummy_strings :: [String] -> NetFunction

dummy_cidr :: CIDR -> NetFunction

dummy_cidrs :: [CIDR] -> NetFunction

crc16 :: NetFunction

steer_gtp_usr :: IPv4 -> Int -> NetFunction

Dispatch the packet across the sockets with a randomized algorithm that maintains the integrity of per-user flows on top of GTP tunnel protocol (Control-Plane packets are broadcasted to all sockets).

(steer_gtp_usr "192.168.0.0" 16)

gtp :: NetFunction

Evaluate to Pass SkBuff in case of GTP packet, Drop it otherwise.

gtp_cp :: NetFunction

Evaluate to Pass SkBuff in case of GTP Control-Plane packet, Drop it otherwise.

gtp_up :: NetFunction

Evaluate to Pass SkBuff in case of GTP User-Plane packet, Drop it otherwise.

is_gtp :: NetPredicate

Evaluate to True if the SkBuff is a GTP packet.

is_gtp_cp :: NetPredicate

Evaluate to True if the SkBuff is a GTP Control-Plane packet.

is_gtp_up :: NetPredicate

Evaluate to True if the SkBuff is a GTP User-Plane packet.

shift :: NetFunction -> NetFunction

trace :: NetFunction

Log monadic/state information to syslog.

udp >-> log_msg "This is an UDP packet"