Network.PFq

Description

The PFq module is a low level binding to the functions in libpfq. See https://github.com/pfq/PFQ/wiki for more information.

Types

data PFqTag

Packet capture handle.

data Statistics

PFq statistics.

Constructors

Statistics

Fields

sReceived :: Integer

packets received

sLost :: Integer

packets lost

sDropped :: Integer

packets dropped

sSent :: Integer

packets sent

sDiscard :: Integer

packets discarded

data NetQueue

Capture Queue handle.

Constructors

NetQueue

Fields

qPtr :: Ptr PktHdr

pointer to the memory mapped queue

qLen :: !Word64

queue length

qSlotSize :: !Word64

size of a slot = pfq header + packet

qIndex :: !Word32

index of the queue

data Packet

Descriptor of the packet.

Constructors

Packet

Fields

pHdr :: Ptr PktHdr

pointer to pfq packet header

pData :: Ptr Word8

pointer to the packet data

pIndex :: !Word32

index of the queue

data PktHdr

PFq packet header.

Constructors

PktHdr

Fields

hData :: !Word64

opaque 64-bits mark

hSec :: !Word32

timestamp (seconds)

hNsec :: !Word32

timestamp (nanoseconds)

hIfIndex :: !Word32

interface index

hGid :: !Word32

group id

hLen :: !Word16

packet length (wire size)

hCapLen :: !Word16

capture length

hTci :: !Word16

vlan tci

hHwQueue :: !Word8

hardware queue index

hCommit :: !Word8

commit bit

type Callback = PktHdr -> Ptr Word8 -> IO ()

The type of the callback function passed to dispatch.

newtype ClassMask

ClassMask type.

Constructors

ClassMask

Fields

getClassMask :: CULong

class_default :: ClassMask

class_any :: ClassMask

newtype GroupPolicy

Group policy type.

Constructors

GroupPolicy

Fields

getGroupPolicy :: CInt

policy_undefined :: GroupPolicy

policy_priv :: GroupPolicy

policy_restricted :: GroupPolicy

policy_shared :: GroupPolicy

newtype PFqConstant

Generic pfq constant.

Constructors

PFqConstant

Fields

getConstant :: Int

Socket and Groups

open

Arguments

:: Int
caplen

-> Int
offset

-> Int
number of rx slots

-> IO (ForeignPtr PFqTag)

Open a socket and join a new group.

The default values for class mask and group policy are class_default and policy_priv, respectively.

openTx

Arguments

:: Int
maxlen

-> Int
number of tx slots

-> IO (ForeignPtr PFqTag)

Open the socket for transmission. No group is joined.

The group policy is policy_undefined by default.

openNoGroup

Arguments

:: Int
caplen

-> Int
offset

-> Int
number of rx slots

-> IO (ForeignPtr PFqTag)

Open a socket. No group is joined.

Groups can later be joined by means of the joinGroup function.

openGroup

Arguments

:: [ClassMask]
list of ClassMask (i.e. [class_default])

-> GroupPolicy
policy for the group

-> Int
caplen

-> Int
offset

-> Int
number of rx slots

-> Int
maxlen

-> Int
number of tx slots

-> IO (ForeignPtr PFqTag)

Open a socket.

Both class mask and group policy are specifiable.

close :: Ptr PFqTag -> IO ()

Close the socket.

enable :: Ptr PFqTag -> IO ()

Enable the socket for packet capture.

disable :: Ptr PFqTag -> IO ()

Disable the socket for packet capture.

getId :: Ptr PFqTag -> IO Int

Return the id of the socket.

getGroupId :: Ptr PFqTag -> IO Int

Return the group id of the socket.

isEnabled :: Ptr PFqTag -> IO Bool

Check whether the packet capture is enabled.

bind

Arguments

:: Ptr PFqTag

-> String
device name

-> Int
queue index (-1 means any queue)

-> IO ()

Bind the main group of the socket to the given device/queue.

bindGroup

Arguments

:: Ptr PFqTag

-> Int
group id

-> String
device name

-> Int
queue index

-> IO ()

Bind the group to the given device/queue.

unbind

Arguments

:: Ptr PFqTag

-> String
device name

-> Int
queue index

-> IO ()

Unbind the main group of the socket from the given device/queue.

unbindGroup

Arguments

:: Ptr PFqTag

-> Int
group id

-> String
device name

-> Int
queue index

-> IO ()

Unbind the group from the given device/queue.

joinGroup

Arguments

:: Ptr PFqTag

-> Int
group id

-> [ClassMask]
class mask list

-> GroupPolicy
group policy

-> IO ()

Join the group with the given class mask and group policy.

leaveGroup

Arguments

:: Ptr PFqTag

-> Int
group id

-> IO ()

Leave the group.

setTimestamp

Arguments

:: Ptr PFqTag

-> Bool
toggle: true is on, false is off.

-> IO ()

Set the timestamping for packets.

getTimestamp :: Ptr PFqTag -> IO Bool

Check whether the timestamping for packets is enabled.

setPromisc

Arguments

:: Ptr PFqTag

-> String
device name

-> Bool
toggle: true is on, false is off.

-> IO ()

Set the promiscuous mode for the given interface.

Packet capture

read

Arguments

:: Ptr PFqTag

-> Int
timeout (msec)

-> IO NetQueue

Read packets in place.

Wait for packets and return a NetQueue descriptor. The timeout is specified in microseconds.

dispatch

Arguments

:: Ptr PFqTag

-> Callback
packet processing function

-> Int
timeout (msec)

-> IO ()

Collect and process packets.

This function is passed a function Callback which is called on each packet.

getPackets :: NetQueue -> IO [Packet]

Return the list of Packet stored in the NetQueue.

getPacketHeader :: Packet -> IO PktHdr

Return the PktHdr of the given the Packet.

isPacketReady :: Packet -> IO Bool

Check whether the Packet is ready or not.

waitForPacket :: Packet -> IO ()

Wait until the Packet is ready.

getCaplen :: Ptr PFqTag -> IO Int

Return the capture length of packets, in bytes.

setCaplen

Arguments

:: Ptr PFqTag

-> Int
caplen (bytes)

-> IO ()

Specify the capture length of packets, in bytes.

Capture length must be set before the socket is enabled for capture.

getOffset :: Ptr PFqTag -> IO Int

Return the capture offset of packets.

setOffset

Arguments

:: Ptr PFqTag

-> Int
offset

-> IO ()

Specify the capture offset of packets, in bytes.

getRxSlots :: Ptr PFqTag -> IO Int

Return the length of the RX queue, in number of packets.

setRxSlots

Arguments

:: Ptr PFqTag

-> Int
number of rx slots

-> IO ()

Specify the length of the RX queue, in number of packets./

The number of RX slots can't exceed the max value specified by the rx_queue_slot kernel module parameter.

getRxSlotSize :: Ptr PFqTag -> IO Int

Return the length of a RX slot, in bytes.

vlanFiltersEnabled

Arguments

:: Ptr PFqTag

-> Int
group id

-> Bool
toggle: true is on, false is off.

-> IO ()

Set the vlan filtering for the given group.

vlanSetFilterId

Arguments

:: Ptr PFqTag

-> Int
group id

-> Int
vlan id

-> IO ()

Specify a capture filter for the given group and vlan id.

vlanResetFilterId

Arguments

:: Ptr PFqTag

-> Int
group id

-> Int
vlan id

-> IO ()

Reset the vlan filter.

Packet transmission

bindTx :: Ptr PFqTag -> String -> Int -> IO ()

Bind the socket for transmission to the given device name and queue.

A socket for transmission can be bound to a given device/queue at time.

startTxThread

Arguments

:: Ptr PFqTag

-> Int
node

-> IO ()

Start the TX kernel thread for transmission.

stopTxThread :: Ptr PFqTag -> IO ()

Stop the TX kernel thread.

wakeupTxThread :: Ptr PFqTag -> IO ()

Wakeup the TX kernel thread.

Wake up the TX kernel thread which transmits the packets in the Tx queue. The kernel thread must be already started.

txQueueFlush :: Ptr PFqTag -> IO ()

Flush the TX queue, in the user context of the calling thread.

To invoke this function, no kernel thread is required.

inject

Arguments

:: Ptr PFqTag

-> ByteString
bytes of packet

-> IO ()

Schedule the packet for transmission.

The packet is copied into the TX queue and sent when the txQueueFlush or the wakeupTxThread function is invoked.

send

Arguments

:: Ptr PFqTag

-> ByteString
bytes of packet

-> IO ()

Transmit the packet passed.

The packet is copied into the TX queue and sent immediately.

sendSync

Arguments

:: Ptr PFqTag

-> ByteString
bytes of packet

-> Int
length of the batch

-> IO ()

Store the packet and possibly transmit the packets in the queue, synchronously.

The transmission is invoked every n packets enqueued.

sendAsync

Arguments

:: Ptr PFqTag

-> ByteString
bytes of packet

-> Int
length of the batch

-> IO ()

Store the packet and possibly transmit the packets in the queue, asynchronously.

The transmission is invoked by the kernel thread, every n packets enqueued.

getTxSlots :: Ptr PFqTag -> IO Int

Return the length of the TX queue, in number of packets.

setTxSlots

Arguments

:: Ptr PFqTag

-> Int
numer of tx slots

-> IO ()

Specify the length of the TX queue, in number of packets.

The number of TX slots can't exceed the max value specified by the tx_queue_slot kernel module parameter.

getMaxlen :: Ptr PFqTag -> IO Int

Return the max transmission length of packets, in bytes.

setMaxlen :: Ptr PFqTag -> Int -> IO ()

Specify the max transmission length of packets, in bytes.

Q-lang

groupComputation

Arguments

:: Ptr PFqTag

-> Int
group id

-> Computation QFunction
computation (PFqLang)

-> IO ()

Specify a functional computation for the given group.

The functional computation is specified by the q-lan eDLS.

Statistics and counters

getStats :: Ptr PFqTag -> IO Statistics

Return the socket statistics.

getGroupStats

Arguments

:: Ptr PFqTag

-> Int
group id

-> IO Statistics

Return the statistics of the given group.

getGroupCounters

Arguments

:: Ptr PFqTag

-> Int
group id

-> IO Counters

Return the set of counters of the given group.