Network.PFq.Experimental

Experimental Functions

This set of experimental functions may be subject to changes in future releases

dummy :: CInt -> NetFunction

dummy_vector :: [CInt] -> NetFunction

dummy_string :: String -> NetFunction

dummy_strings :: [String] -> NetFunction

crc16 :: NetFunction

class' :: CInt -> NetFunction

Specify the class mask for the given packet.

deliver :: CInt -> NetFunction

par3 :: NetFunction -> NetFunction -> NetFunction -> NetFunction

Function that returns the parallel of 3 monadic NetFunctions.

par4 :: NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction

par5 :: NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction

par6 :: NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction

par7 :: NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction

par8 :: NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction -> NetFunction

steer_gtp_usr :: IPv4 -> CInt -> NetFunction

Dispatch the packet across the sockets with a randomized algorithm that maintains the integrity of per-user flows on top of GTP tunnel protocol (Control-Plane packets are broadcasted to all sockets).

(steer_gtp_usr "192.168.0.0" 16)

gtp :: NetFunction

Evaluate to Pass SkBuff in case of GTP packet, Drop it otherwise.

gtp_cp :: NetFunction

Evaluate to Pass SkBuff in case of GTP Control-Plane packet, Drop it otherwise.

gtp_up :: NetFunction

Evaluate to Pass SkBuff in case of GTP User-Plane packet, Drop it otherwise.

is_gtp :: NetPredicate

Evaluate to True if the SkBuff is a GTP packet.

is_gtp_cp :: NetPredicate

Evaluate to True if the SkBuff is a GTP Control-Plane packet.

is_gtp_up :: NetPredicate

Evaluate to True if the SkBuff is a GTP User-Plane packet.