Network.PFq.Lang

Basic types

newtype IPv4

IPv4 data type

Constructors

IPv4

Fields

getIP4Address :: HostAddress

data Argument

Argument data type. Any PFQ/lang function can take up to 8 Arguments.

Constructors

forall a . (Show a, Storable a) => ArgData a

forall a . (Show a, Storable a) => ArgVector [a]

ArgString String

ArgSVector [String]

ArgFunPtr Int

ArgNull

class Pretty x where

Pretty class, typeclass used to print a PFQ/lang computation.

Methods

pretty :: x -> String

data Function fun where

Parametric Function data type.

Constructors

MFunction :: forall a b c d e f g h. (Serializable a, Argumentable a, Serializable b, Argumentable b, Serializable c, Argumentable c, Serializable d, Argumentable d, Serializable e, Argumentable e, Serializable f, Argumentable f, Serializable g, Argumentable g, Serializable h, Argumentable h) => Symbol -> a -> b -> c -> d -> e -> f -> g -> h -> NetFunction

Predicate :: forall a b c d e f g h. (Serializable a, Argumentable a, Serializable b, Argumentable b, Serializable c, Argumentable c, Serializable d, Argumentable d, Serializable e, Argumentable e, Serializable f, Argumentable f, Serializable g, Argumentable g, Serializable h, Argumentable h) => Symbol -> a -> b -> c -> d -> e -> f -> g -> h -> NetPredicate

Property :: forall a b c d e f g h. (Serializable a, Argumentable a, Serializable b, Argumentable b, Serializable c, Argumentable c, Serializable d, Argumentable d, Serializable e, Argumentable e, Serializable f, Argumentable f, Serializable g, Argumentable g, Serializable h, Argumentable h) => Symbol -> a -> b -> c -> d -> e -> f -> g -> h -> NetProperty

Combinator1 :: Symbol -> NetPredicate -> NetPredicate

Combinator2 :: Symbol -> NetPredicate -> NetPredicate -> NetPredicate

Composition :: forall f1 f2 f. (Serializable (Function f1), Serializable (Function f2)) => Function f1 -> Function f2 -> Function f

class Serializable a where

Serializable class, a typeclass used to serialize computations. Transform a Function into a list of FunctionDescr.

Methods

serialize :: a -> Int -> ([FunctionDescr], Int)

data FunctionDescr

Function descriptor.

Constructors

FunctionDescr

Fields

functionSymbol :: Symbol

functionArguments :: [Argument]

functionIndex :: Int

functionLink :: Int

data Action a

Action is a monad modelled after the Identity and implemented at kernel level.

data SkBuff

SkBuff placeholder type is used to model the kernel sk_buff data structure.

Function types

type NetFunction = Function (SkBuff -> Action SkBuff)

Simple monadic in-kernel PFQ/lang function.

type NetPredicate = Function (SkBuff -> Bool)

Simple in-kernel PFQ/lang predicate.

type NetProperty = Function (SkBuff -> Word64)

Simple in-kernel PFQ/lang property function.

(>->) :: Function (a -> m b) -> Function (b -> m c) -> Function (a -> m c)

Kleisli left-to-right operator, for monadic composition of PFQ/lang functions.