Network.PFQ.Lang

Basic types

newtype IPv4

IPv4 data type

Constructors

IPv4

Fields

getHostAddress :: HostAddress

newtype CIDR

CIDR data-type

Constructors

CIDR

Fields

getNetworkPair :: (IPv4, Int)

data Argument

Argument data type. Any pfq-lang function can take up to 8 Arguments.

Constructors

forall a . (Show a, Storable a, Typeable a, ToJSON a, FromJSON a) => ArgData a

forall a . (Show a, Storable a, Typeable a, ToJSON a, FromJSON a) => ArgVector [a]

ArgString String

ArgStrings [String]

ArgFunPtr Int

ArgNull

class Pretty x where

Pretty class, typeclass used to print a pfq-lang computation.

Methods

pretty :: x -> String

data Function fun where

Parametric Function data type.

Constructors

Function :: forall a b c d e f g h. (ArgumentClass a, ArgumentClass b, ArgumentClass c, ArgumentClass d, ArgumentClass e, ArgumentClass f, ArgumentClass g, ArgumentClass h) => Symbol -> a -> b -> c -> d -> e -> f -> g -> h -> NetFunction

Predicate :: forall a b c d e f g h. (ArgumentClass a, ArgumentClass b, ArgumentClass c, ArgumentClass d, ArgumentClass e, ArgumentClass f, ArgumentClass g, ArgumentClass h) => Symbol -> a -> b -> c -> d -> e -> f -> g -> h -> NetPredicate

Property :: forall a b c d e f g h. (ArgumentClass a, ArgumentClass b, ArgumentClass c, ArgumentClass d, ArgumentClass e, ArgumentClass f, ArgumentClass g, ArgumentClass h) => Symbol -> a -> b -> c -> d -> e -> f -> g -> h -> NetProperty

Combinator1 :: Symbol -> NetPredicate -> NetPredicate

Combinator2 :: Symbol -> NetPredicate -> NetPredicate -> NetPredicate

Kleisli :: forall f1 f2 f. (Serializable (Function f1), Serializable (Function f2)) => Function f1 -> Function f2 -> Function f

class Serializable a where

Serializable class, a typeclass used to serialize computations. Transform a Function into a list of FunctionDescr.

Methods

serialize :: a -> Int -> ([FunctionDescr], Int)

data FunctionDescr

Function descriptor.

Constructors

FunctionDescr

Fields

funSymbol :: Symbol

funArgs :: [Argument]

funIndex :: Int

funLink :: Int

newtype SkBuff

SkBuff placeholder type is used to model the kernel sk_buff data structure.

Constructors

SkBuff ()

data Action a

Action is a monad modelled after the Identity as it is implemented at kernel level.

Function types

type NetFunction = Function (SkBuff -> Action SkBuff)

Simple monadic in-kernel pfq-lang function.

type NetPredicate = Function (SkBuff -> Bool)

Simple in-kernel pfq-lang predicate.

type NetProperty = Function (SkBuff -> Word64)

Simple in-kernel pfq-lang property function.

(>->) :: forall a b c m. Monad m => Function (a -> m b) -> Function (b -> m c) -> Function (a -> m c)

Kleisli left-to-right operator